requirements.txt
Django==1.9.5
whitenoise==3.1
djangorestframework==3.3.3
Markdown==2.6.6
django-filter==0.13.0